+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## CHANGELOG
 
-### 2.5.0 (15-Oct-2025)
+### 2.5.0 (16-Oct-2025)
 - Added "System Memory" and "System Storage" capacity information (Pull Request #36; thanks again, @HowardGMac!)
 - Corrected misspelling of "Certificate" in multiple locations (Pull Request #41; thanks, @HowardGMac!)
 - Improved handling of the `checkJamfProCheckIn` and `checkJamfProInventory` functions when no relevant data is found in the `jamf.log` file
@@ -13,6 +13,9 @@
 - Added the size and item count of the user's Desktop and Trash to the Jamf Pro Policy Log Reporting
 - Added `checkUserDirectorySizeItems` function to report the size and item count of any user directories (e.g. Desktop, Downloads, Trash, etc.)
 - Added a Health Checks for Signed System Volume (SSV) and Gatekeeper / XProtect (thanks for the reminder, @hoakley!)
+- Refactored "DDM-enforced OS Version" per [DDM-OS-Reminder](https://github.com/dan-snelson/DDM-OS-Reminder)
+- Refactored `checkUserDirectorySizeItems` to ignore hidden files
+- Simplified various date / time formats
 - Introduces a new `operationMode` of "Silent" to run all checks and log results without displaying a dialog to the end-user
 
     > :warning: **Breaking Change** :warning: